Add `math.atan`, arc tangent function, and test
Add a continued-fraction approximation of the arc tangent function
for real-valued arguments. The implementation follows the construct in
https://en.wikipedia.org/wiki/Inverse_trigonometric_functions .

Also add a unit test that checks basic consistency by verifying that
the atan of the tan of an argument is the argument again; same for
tan of atan, and negative arguments to check symmetries.