XXX Work around `apt` authentication error in Travis
THIS IS A POTENTIALLY DANGEROUS HOTFIX! DO NOT USE IN PRODUCTION!

There is a package installation problem in Travis, one of the Continuous
Integration systems we use, for Linux with Python 2.5 via @deadsnakes's /
@fkrull's Private Package Archive (PPA). Specifically, packages that used
to e authenticated correctly now fail to authenticate
(see https://travis-ci.org/aaaaalbert/seattlelib_v2/jobs/271635113#L524
and https://travis-ci.org/aaaaalbert/seattlelib_v2/jobs/355737211#L497 ).

This POTENTIALLY DANGEROUS HOTFIX allows unauthenticated packages to be
installed so that I can temporarily run Python 2.5 tests again. However,
this hotfix should not be used in production, at least not until we
understand the problem and the correct ways to fix it better.